backend/requirements.txt
asgiref==3.8.1
astroid==3.3.5
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.4
dill==0.3.9
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-phonenumber-field==7.3.0
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.48
Pillow==10.1.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.2
urllib3==2.2.3
